[SECURESIGN-2748] Add test: simulate Rekor server unavailability (#26)
* add test with empty Rekor log

* review updates

* refactor: Get rid of asynchronous eventing in test

* small clean-up

* update deps

---------
